lambda-functions/requirements.txt
boto3==1.34.0
botocore==1.34.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0